Don't handle health-check messages in fallback chain
If the fallback chain takes more than 2 seconds to resolve a query
the fallback plugin's internal forward plugin will think it has failed
and begin sending health checks every 0.5 seconds.

By immediately returning any reply for the healthcheck message
(even an error), it will be deemed a healthy upstream server, and
health-checks will stop.

If Cloudflare is inaccessible then it is very easy for it to take more
than 2 seconds to process a request, and piling on a bunch of
health-check messages makes things much much worse.

Refusing this specific request should have no ill effect, as there
is not normally any need to make an `IN NS .` request, especially
of an stub or recursive resolver, and in any event, normally the main
chain would handle this.