Use CoreDNS default cache settings
Currently our cache settings limit the maximum time a DNS record can be
cached to 600 seconds. However, from code history it is unclear why
cache is set to that as maximum.

A DNS entry which changes often has a lower TTL in it's response, which
will be still accepted. So using the default cache setting only affects
records which do have a higher TTL on their end already (and thuse
likely anyways cached in the usptream DNS).